+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 0.2.0
+- Added x86 Page Table Entry Visualizer tool
+- Footer version is now a button; clicking will show current version's changelog
 - Added a toggle button to switch between two-column and single-column layout modes
-- Footer version is now a button: clicking it will show current version's changelog
 
 # 0.1.0
 - Added x86 Address Converter tool